upload exceptions

+++ src/main/java/br/ufpb/dcx/lab1v1/controllers/DisciplinaController.java
@@ -2,6 +2,8 @@ package br.ufpb.dcx.lab1v1.controllers;
 
 import br.ufpb.dcx.lab1v1.dtos.DisciplinaDTO;
 import br.ufpb.dcx.lab1v1.entidades.Disciplina;
+import br.ufpb.dcx.lab1v1.excecoes.DisciplinaInvalidaException;
+import br.ufpb.dcx.lab1v1.excecoes.DisciplinaJaExistenteException;
 import br.ufpb.dcx.lab1v1.services.DisciplinaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -31,27 +33,27 @@ public class DisciplinaController {
     }
 
     @GetMapping("/v1/api/disciplinas/{id}")
-    public ResponseEntity<Disciplina> retornaDisciplina(@PathVariable int id) {
+    public ResponseEntity<Disciplina> retornaDisciplina(@PathVariable Long id) {
         return ResponseEntity.status(HttpStatus.OK).body(disciplinaService.retornaDisciplina(id));
     }
 
     @PostMapping("/v1/api/disciplinas/{id}/nome")
-    public ResponseEntity<Disciplina> atualizaDisciplina(@PathVariable int id, @RequestBody DisciplinaDTO disciplinaDTO) {
+    public ResponseEntity<Disciplina> atualizaDisciplina(@PathVariable Long id, @RequestBody DisciplinaDTO disciplinaDTO) {
         return ResponseEntity.status(HttpStatus.OK).body(disciplinaService.atualizaDisciplina(id, disciplinaDTO.getNome()));
     }
 
     @PostMapping("/v1/api/disciplinas/{id}/nota")
-    public ResponseEntity<Disciplina> adicionaNota(@PathVariable int id, @RequestBody DisciplinaDTO disciplinaDTO) {
+    public ResponseEntity<Disciplina> adicionaNota(@PathVariable Long id, @RequestBody DisciplinaDTO disciplinaDTO) {
         return ResponseEntity.status(HttpStatus.OK).body(disciplinaService.adicionaNota(id, disciplinaDTO.getNota()));
     }
 
     @PostMapping("/v1/api/disciplinas/{id}/like")
-    public ResponseEntity<Disciplina> adicionaLike(@PathVariable int id) {
+    public ResponseEntity<Disciplina> adicionaLike(@PathVariable Long id) {
         return ResponseEntity.status(HttpStatus.OK).body(disciplinaService.adicionaLike(id));
     }
 
     @DeleteMapping("/v1/api/disciplinas/{id}")
-    public ResponseEntity<Disciplina> removeDisciplina(@PathVariable int id) {
+    public ResponseEntity<Disciplina> removeDisciplina(@PathVariable Long id) {
         return ResponseEntity.status(HttpStatus.OK).body(disciplinaService.removeDisciplina(id));
     }
