+++ src/main/java/br/ufpb/dcx/lab1v1/entidades/Disciplina.java
@@ -1,12 +1,17 @@
 package br.ufpb.dcx.lab1v1.entidades;
 
+import br.ufpb.dcx.lab1v1.excecoes.DisciplinaInvalidaException;
+import br.ufpb.dcx.lab1v1.excecoes.DisciplinaJaExistenteException;
+import br.ufpb.dcx.lab1v1.excecoes.IdInvalidoException;
+import br.ufpb.dcx.lab1v1.excecoes.NotaInvalidaException;
+
 import java.util.ArrayList;
 import java.util.List;
 
 public class Disciplina implements Comparable<Disciplina>{
 
-    private static int proximoID = 1;
-    private int id;
+    private static long proximoID = 1;
+    private long id;
     private String nome;
     private int likes;
     private List<Double> notas = new ArrayList<>();
@@ -44,7 +49,7 @@ public class Disciplina implements Comparable<Disciplina>{
         this.notas = notas;
     }*/
 
-    public void adicionarNota(double nota) {
+    public void adicionarNota(Double nota) {
         notas.add(nota);
     }
